+ Refactoring CommentForm  + Adding comments to self-care techniques
