+++ pages/self-care/[technique].js
@@ -1,20 +1,32 @@
 import styled from 'styled-components'
-import { useState } from 'react'
-
-import client from '../../client'
 
 import Container from '../../components/Container'
+import Comments from '../../components/Comment'
+import CommentForm from '../../components/CommentForm'
 import Block from '../../components/Block'
+
+import client from '../../client'
 
 const GET_SELF_CARE_TECHNIQUE = `*[_type == "selfcareTechnique" && slug.current == $slug][0]{
   _type,
   title,
-  mainImage,
-  "videoUrl": video.asset->url,
-  body,
-  audio, 
-  link, 
-  likes
+  "technique": {
+    "parentID": _id,
+    title,
+    "videoUrl": video.asset->url,
+    body,
+    audio, 
+    link, 
+    likes,
+    mainImage,
+  },
+  "form": *[_type == "form" && slug.current == "selfcare-technique-comment"][0]{
+    "formTitle": title, 
+    body, 
+    subtitle, 
+    "inputsFromSanity": inputs[]->{ title, required, type },
+    confirmationText
+  }
 }`
 
 const TechniqueStyled = styled(Container).attrs({
@@ -33,9 +45,11 @@ const OuterContainer = styled.div.attrs({
   className: 'bg-white min-h-content',
 })``
 
-const Technique = props => {
-  const [technique] = useState(props)
-
+const Technique = ({
+  query: { technique: slug },
+  technique,
+  form: { formTitle, subtitle, inputsFromSanity, confirmationText },
+}) => {
   return (
     <OuterContainer className="h-full bg-white">
       {technique.videoUrl && (
@@ -47,8 +61,18 @@ const Technique = props => {
           body={technique.body}
           className="font-sm font-normal"
           imageOptions={{ w: 320, h: 240, fit: 'max' }}
-        />{' '}
+        />
       </TechniqueStyled>
+      <CommentForm
+        {...{
+          inputsFromSanity,
+          title: formTitle,
+          subtitle,
+          confirmationText,
+          parentID: technique.parentID,
+        }}
+      />
+      <Comments {...{ slug, _type: 'selfcareTechnique' }} />
     </OuterContainer>
   )
 }
