+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){const e=e=>{"require"!==e&&(e+=".js");let a=Promise.resolve();return c[e]||(a=new Promise(async a=>{if("document"in self){const c=document.createElement("script");c.src=e,document.head.appendChild(c),c.onload=a}else importScripts(e),a()})),a.then(()=>{if(!c[e])throw new Error(`Module ${e} didn’t register its module`);return c[e]})},a=(a,c)=>{Promise.all(a.map(e)).then(e=>c(1===e.length?e[0]:e))},c={require:Promise.resolve(a)};self.define=(a,s,i)=>{c[a]||(c[a]=Promise.resolve().then(()=>{let c={};const r={uri:location.origin+a.slice(1)};return Promise.all(s.map(a=>{switch(a){case"exports":return c;case"module":return r;default:return e(a)}})).then(e=>{const a=i(...e);return c.default||(c.default=a),c})}))}}define("./sw.js",["./workbox-e032be30"],(function(e){"use strict";importScripts(),e.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/",revision:"iGAXNpsLVTmP6oMyXNxY6"},{url:"/_next/static/chunks/2f86b3c7615ad4e99d0e9e89c93bb08f45ef196e.094860d52acdcd6c9b0a.js",revision:"45b7d0059f987c962ba2f049e463367e"},{url:"/_next/static/chunks/36d8bf268fddc9c16e55909be2f58ae43ed726dc.26418a335cf7101c4d63.js",revision:"9da53ce70826f27a94fdabc9daf3ec3c"},{url:"/_next/static/chunks/3e5088394defbd1b84c5306afde0bf61cc90c18e.663a53ea5e0591fe6753.js",revision:"8d9d76722fcacd2c0a146a2046f77f94"},{url:"/_next/static/chunks/427c0cb2593e776d698af642f8d33ed525d05213.852fb146efcafc0bf77b.js",revision:"f1ed9a8e70ee5c441d51d4bba540007c"},{url:"/_next/static/chunks/42e3a096f2fc1c2957c7f309a89226b5d3f63a0e.b894c8bc7d2d6f749948.js",revision:"d88f317f1e6abb8861f05ad5c6820ffd"},{url:"/_next/static/chunks/50e7ebedd171f1912375c1fd7a19d8905184f73c.c70bc0c2ecf16ec63af9.js",revision:"627f68de1e85c40a449280e21f4094d7"},{url:"/_next/static/chunks/73795c3f.a6fdd85086dc91ce02c7.js",revision:"0b43f1ba60a9cbe49ed3254c8144b149"},{url:"/_next/static/chunks/75fc9c18.090c03733b865a37e886.js",revision:"8e37215134f954b77c15f5f4c5b9da05"},{url:"/_next/static/chunks/88ae3f376e6ece4e043b2297d1b67309db1be995.9f10cce45e3c16d45009.js",revision:"edf5b27b4d54d93aaf0878eb8209f40e"},{url:"/_next/static/chunks/a3c50e9434aaec8cb791c2fdd58942c1460d6f3b.1377f28eb8968ae15bb6.js",revision:"5ce9ff5c4ff437f1d1e37ce098d4addd"},{url:"/_next/static/chunks/a969f3c36a4d4dd1b718cd8fc871f448358c3d75.b0a46419e66eceef3963.js",revision:"1fe5ecb4e4fdd79e4811c3b0bb5b010a"},{url:"/_next/static/chunks/c52d967564a3b93fde5371e3457b5bf1be5e1171.a5cd44af2758bb15f634.js",revision:"3d3e525f987538e2a78ba22e39b99fd8"},{url:"/_next/static/chunks/f3e33abbadae7ac015df44189a8115bf3b4370d0.caf4b180e968a1307815.js",revision:"73a05ed5bbe4028fc7f1827d551311a4"},{url:"/_next/static/chunks/framework.4dd1003cc9c949c7fcd3.js",revision:"8dbfd54516c12914d3e0cd417cd67882"},{url:"/_next/static/chunks/styles.daa31c95f3f4d471c1e2.js",revision:"bf4bc896d73c22d9cf9edc543584cb0f"},{url:"/_next/static/css/73795c3f.effeec63.chunk.css",revision:"35c05ee9fe280a818dd708415c2473e2"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/_buildManifest.js",revision:"20b39f75ec7946c95a77d67fa856f7fa"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/_ssgManifest.js",revision:"abee47769bf307639ace4945f9cfd4ff"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/_apollo_example.js",revision:"c7d365afbb746c18df34d65b5d7fc415"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/_app.js",revision:"72101b9b3697ce3de010a4b28fd56005"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/_error.js",revision:"479f79936c102817eba367fc75de156d"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/about-ava.js",revision:"033ad292296576ef71dc2a55de674405"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/add-to-homescreen.js",revision:"31236084fc5c4f724bbda652a15fd293"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/am-i-in-an-abusive-relationship.js",revision:"312dcc422cd3fe54205259a501c54941"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/contact-us.js",revision:"7fd3e73d7c0e29399f280a3d65ab1de3"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/cookies.js",revision:"89f9a91e9f0428999f76485feb4e3fbb"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/index.js",revision:"8f0d7cc19a09800e2ad451ec6d19ae64"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/landing.js",revision:"01a89509bc046db779eeacd5e4fbf130"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/onboarding.js",revision:"45731b4995a7078d0216a8292ca265dc"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/privacy.js",revision:"d1f62e72e46f151afaf108bb733f0bcc"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/self-care.js",revision:"305ca3ab2aa958dbcd00b6638115f4e2"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/self-care/%5Btechnique%5D.js",revision:"543e9eb6ac0564bafa19698eb8d799f0"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/stories.js",revision:"d191c1a9b8089fa9dd42ce84381796f4"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/stories/%5Bstory%5D.js",revision:"7bb6b15dcbd21f644b7cdfe8cbf63397"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/stories/share-your-story.js",revision:"108375b3015c499ac7a8b7abfc8c5a36"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/support/%5Bcategory%5D.js",revision:"18d969303a26cbc5e828b1f0c50a9ca3"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/support/%5Bcategory]/%5Bservice%5D.js",revision:"7d929d0eb3341e1697efc52e5ca6ca22"},{url:"/_next/static/iGAXNpsLVTmP6oMyXNxY6/pages/support/%5Bcategory]/filter.js",revision:"e8a33cc50e01134e6c2693c4602b6885"},{url:"/_next/static/runtime/main-df269a212949917dc87a.js",revision:"505b8a5815ea73a45f8675bb39eb4f51"},{url:"/_next/static/runtime/polyfills-0e7ae36d1087531904bf.js",revision:"9dfafb5d1d1b545064299aa28494fe0f"},{url:"/_next/static/runtime/webpack-83bd83ab777f80a6c75c.js",revision:"f5e6e2fca3144cc944812cfa3547f475"},{url:"/android-chrome-192x192.png",revision:"c87b3ef459c3d9e2ca89a7ce214ea64d"},{url:"/android-chrome-512x512.png",revision:"efa98c7406efc5721bade2c73ffb7ec5"},{url:"/apple-touch-icon.png",revision:"875608d0dc40d89e9b48b94e82489103"},{url:"/favicon-16x16.png",revision:"c86804fcb5629d4b5d5e8099439d9b7f"},{url:"/favicon-32x32.png",revision:"0cbcefe245f1bdfed30f1b48f8351ce6"},{url:"/favicon.ico",revision:"412192267449ea67eebabd3e62acfe51"},{url:"/fonts/caladea-v1-latin-regular.eot",revision:"8c187d3ae793150ec4c8054145080d49"},{url:"/fonts/caladea-v1-latin-regular.svg",revision:"eafdfaa360fef8bd125a93ffa3249499"},{url:"/fonts/caladea-v1-latin-regular.ttf",revision:"77ecd4059bfe360cede8f1b02b456404"},{url:"/fonts/caladea-v1-latin-regular.woff",revision:"ec24dc571a9b42b75a902dcd8ea9f038"},{url:"/fonts/caladea-v1-latin-regular.woff2",revision:"0610fc99421e0e6ca4c4c7c14ae2f43d"},{url:"/fonts/karla-v13-latin-700.eot",revision:"9c56478b8cf084983b407c96f657a2d1"},{url:"/fonts/karla-v13-latin-700.svg",revision:"a2332f3209bfe951084c20133c2e5a5d"},{url:"/fonts/karla-v13-latin-700.ttf",revision:"561294dc58294f56b0983bb5619d93da"},{url:"/fonts/karla-v13-latin-700.woff",revision:"a44f3c2705f046cd0afac4bfc2fe1eed"},{url:"/fonts/karla-v13-latin-700.woff2",revision:"57214306eebeea47c94e695a778c3f7b"},{url:"/fonts/karla-v13-latin-regular.eot",revision:"84572d5ccb359b1ae5481312e8db43ab"},{url:"/fonts/karla-v13-latin-regular.svg",revision:"581de7ca06f14fd645f791c8a283be71"},{url:"/fonts/karla-v13-latin-regular.ttf",revision:"61c4f9d8015142b069ab5e7f838594b4"},{url:"/fonts/karla-v13-latin-regular.woff",revision:"1ba33869ba2e0b430fb9cdc948dd0d54"},{url:"/fonts/karla-v13-latin-regular.woff2",revision:"c4aa4ca22f0c3e52b9e4593f63ca57a7"},{url:"/fonts/porpara-regular.ttf",revision:"49b774e03bef78a8bf8942a43fd363c9"},{url:"/icons/arrow.svg",revision:"c80dbb2c7c8ef269eb1da98c43b932f9"},{url:"/icons/back-black.svg",revision:"ab5a3aa611e11274a50ef988d38b09e5"},{url:"/icons/clock.svg",revision:"abde8858933f38dbca2c682c8499df80"},{url:"/icons/email.svg",revision:"c721c9873dcdb77b8b855ee0081a1680"},{url:"/icons/exit-white.svg",revision:"8487fd73fa6cb6e7fbf687fd9a845a50"},{url:"/icons/forwardArrow.svg",revision:"e230f09c9e2bc1337ef785dfbbcee2a2"},{url:"/icons/heart-black.svg",revision:"58370e8d994fd4619735f587e1a6e1ba"},{url:"/icons/icon-128x128.png",revision:"db7481a04968b3ad4716436d30b3e409"},{url:"/icons/icon-144x144.png",revision:"689a1926530ed6ca26db537029c1dfd7"},{url:"/icons/icon-152x152.png",revision:"98af072da209b3b297308717b06ea86c"},{url:"/icons/icon-192x192.png",revision:"7ab8c82581552c7558c5b009ffcb3cd2"},{url:"/icons/icon-384x384.png",revision:"aae6d1f552c78e06bf42aae2462a98fa"},{url:"/icons/icon-512x512.png",revision:"837ab4469c705c188326cec571ecb227"},{url:"/icons/icon-72x72.png",revision:"cb6e3fb9db7db52f0986da1ff9bab188"},{url:"/icons/icon-96x96.png",revision:"7d8348b60cdd985bd24b48cb56855c1a"},{url:"/icons/info.svg",revision:"bc2c7b826d1d41884dcd2e224d33f01b"},{url:"/icons/phone.svg",revision:"6854b54332adb9d004a730dc6fecd95e"},{url:"/icons/right-arrow.svg",revision:"1de339a612f220fda976721291a7a5ee"},{url:"/icons/url.svg",revision:"de9d153bbe15ab7d7121eef87bf12ca8"},{url:"/manifest.json",revision:"688128be216cc2f753fe641590f2fcd3"},{url:"/site.webmanifest",revision:"86e7fdaa2bd5af2341bbe0f006f05576"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute(/^https:\/\/fonts\.(?:googleapis|gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/^https:\/\/use\.fontawesome\.com\/releases\/.*/i,new e.CacheFirst({cacheName:"font-awesome",plugins:[new e.ExpirationPlugin({maxEntries:1,maxAgeSeconds:31536e3,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.StaleWhileRevalidate({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/.*/i,new e.StaleWhileRevalidate({cacheName:"others",plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET")}));
+if(!self.define){const e=e=>{"require"!==e&&(e+=".js");let a=Promise.resolve();return s[e]||(a=new Promise(async a=>{if("document"in self){const s=document.createElement("script");s.src=e,document.head.appendChild(s),s.onload=a}else importScripts(e),a()})),a.then(()=>{if(!s[e])throw new Error(`Module ${e} didn’t register its module`);return s[e]})},a=(a,s)=>{Promise.all(a.map(e)).then(e=>s(1===e.length?e[0]:e))},s={require:Promise.resolve(a)};self.define=(a,c,i)=>{s[a]||(s[a]=Promise.resolve().then(()=>{let s={};const n={uri:location.origin+a.slice(1)};return Promise.all(c.map(a=>{switch(a){case"exports":return s;case"module":return n;default:return e(a)}})).then(e=>{const a=i(...e);return s.default||(s.default=a),s})}))}}define("./sw.js",["./workbox-e032be30"],(function(e){"use strict";importScripts(),e.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/",revision:"W3LbsHiTCaIPQ9Cne_PEf"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/_buildManifest.js",revision:"1424f1d00ccb36a66c9e6f04b7b98561"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/_ssgManifest.js",revision:"abee47769bf307639ace4945f9cfd4ff"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/_apollo_example.js",revision:"45c07ebd625db0c93d587d2bf0fdd58b"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/_app.js",revision:"b7fa0181e7ce470c91e077898624b2e4"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/_error.js",revision:"33bffbbc4997cc08269b1a552564807f"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/about-ava.js",revision:"4982762a4488a2a6b19ae9d29d5756bd"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/add-to-homescreen.js",revision:"3b020f3e1689f032bd4319a1b4f44462"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/am-i-in-an-abusive-relationship.js",revision:"df3646a9494ca5b38b47b1b7c360c6a0"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/contact-us.js",revision:"aa84a6f25fe59eafb55783b9a3c5111c"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/cookies.js",revision:"c98ec41a9c356ac2bd9d3dd15b82e00f"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/index.js",revision:"66e3019ac29cff61035bdbd34c73e07f"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/landing.js",revision:"3e57ea6e42cd52eb24d3abd4fb8822c5"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/onboarding.js",revision:"94aad444a7fa54c0001ea78021f9b6c1"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/privacy.js",revision:"c4773450bcdf410628a5b6223433e931"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/self-care.js",revision:"18fc296e64babc4eb9b29d04572b7101"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/self-care/%5Btechnique%5D.js",revision:"c7874e730223023489572d065b968328"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/stories.js",revision:"3cb82901ce794697f0bb7e12fdd9806b"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/stories/%5Bstory%5D.js",revision:"2e5f73425a6033c417e0ed50c91b5e78"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/stories/share-your-story.js",revision:"2817dcd0b5e88350ee2ba75eab6b70b5"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/support/%5Bcategory%5D.js",revision:"452ba49629b8b0cfeb6f98848807d28d"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/support/%5Bcategory]/%5Bservice%5D.js",revision:"33f3c39adabd9da23d59067ced302002"},{url:"/_next/static/W3LbsHiTCaIPQ9Cne_PEf/pages/support/%5Bcategory]/filter.js",revision:"60516c62d19b276a23686403b75f7178"},{url:"/_next/static/chunks/164d8d8687a4e631cc9e3d59ab5508016fbc9dad.03c2554bc751cb9a6bd5.js",revision:"678d171d466a62064b95deaa13ec4ca6"},{url:"/_next/static/chunks/17a5b2186b2c321aab84edcb0009f5564e749bd0.1df291df914ab8c82630.js",revision:"8d9d76722fcacd2c0a146a2046f77f94"},{url:"/_next/static/chunks/28c6bed0eaa2c24a0d0d576002b1eb41d417dbf1.9f10cce45e3c16d45009.js",revision:"edf5b27b4d54d93aaf0878eb8209f40e"},{url:"/_next/static/chunks/3cb07a65bfa9265dd301abb4040693477ef88b7f.b0a46419e66eceef3963.js",revision:"1fe5ecb4e4fdd79e4811c3b0bb5b010a"},{url:"/_next/static/chunks/73795c3f.09128eed5fe12bd5a1c2.js",revision:"a079ea4c49a0fad6e2213ac51a95d216"},{url:"/_next/static/chunks/75fc9c18.3d38bcf24184396f0112.js",revision:"f37f29e64e1858974fb7454ba89858a7"},{url:"/_next/static/chunks/94b0396b57f89bc1e7d1d5891638a365c2658c14.2655955f5395b37bb04d.js",revision:"73a05ed5bbe4028fc7f1827d551311a4"},{url:"/_next/static/chunks/9536dd2556109ccd0d67f1687b88c724265818e8.852fb146efcafc0bf77b.js",revision:"f1ed9a8e70ee5c441d51d4bba540007c"},{url:"/_next/static/chunks/b091334c7150542c888db8bf3bf7d9bcae4d3f13.659327c22e405c793d9c.js",revision:"d88f317f1e6abb8861f05ad5c6820ffd"},{url:"/_next/static/chunks/c9410842b4a08f50b491bc51826ee56cfcfc974c.c70bc0c2ecf16ec63af9.js",revision:"627f68de1e85c40a449280e21f4094d7"},{url:"/_next/static/chunks/c9f4df360b8b667f173c16ca09fa75226373a6f8.3a240c170a3e1a266807.js",revision:"bcf0113e1fed3575daf96eb016f21146"},{url:"/_next/static/chunks/d34482a3c7ee5e7e53d1898a0e54e592baeda10c.f749f794a29910d14622.js",revision:"9051dbecb4c7185d8d1a522cd784ea96"},{url:"/_next/static/chunks/d7d04403076abce21819cd005c8bf36e920da623.b875002082c4d7ac5abe.js",revision:"5581409ff44da5c122c3840740283c4c"},{url:"/_next/static/chunks/ed4ac8dced2c1392964961064a05cc3257100857.4429c8fdeb20833fb90c.js",revision:"00590349e348563b7ee25c8403b60f02"},{url:"/_next/static/chunks/framework.4dd1003cc9c949c7fcd3.js",revision:"8dbfd54516c12914d3e0cd417cd67882"},{url:"/_next/static/chunks/styles.eb75cfd706118ee2dd1b.js",revision:"e9e334b11632efdef09d23b019c2a7bb"},{url:"/_next/static/css/73795c3f.effeec63.chunk.css",revision:"35c05ee9fe280a818dd708415c2473e2"},{url:"/_next/static/runtime/main-489fb5a8d04157ed2283.js",revision:"015d941060c6dd71d426eceeab5d6bde"},{url:"/_next/static/runtime/polyfills-94218a8b4426e8ee3b56.js",revision:"d7315dc21359cfc95cf31870dac3c9c4"},{url:"/_next/static/runtime/webpack-83bd83ab777f80a6c75c.js",revision:"f5e6e2fca3144cc944812cfa3547f475"},{url:"/android-chrome-192x192.png",revision:"c87b3ef459c3d9e2ca89a7ce214ea64d"},{url:"/android-chrome-512x512.png",revision:"efa98c7406efc5721bade2c73ffb7ec5"},{url:"/apple-touch-icon.png",revision:"875608d0dc40d89e9b48b94e82489103"},{url:"/favicon-16x16.png",revision:"c86804fcb5629d4b5d5e8099439d9b7f"},{url:"/favicon-32x32.png",revision:"0cbcefe245f1bdfed30f1b48f8351ce6"},{url:"/favicon.ico",revision:"412192267449ea67eebabd3e62acfe51"},{url:"/fonts/caladea-v1-latin-regular.eot",revision:"8c187d3ae793150ec4c8054145080d49"},{url:"/fonts/caladea-v1-latin-regular.svg",revision:"eafdfaa360fef8bd125a93ffa3249499"},{url:"/fonts/caladea-v1-latin-regular.ttf",revision:"77ecd4059bfe360cede8f1b02b456404"},{url:"/fonts/caladea-v1-latin-regular.woff",revision:"ec24dc571a9b42b75a902dcd8ea9f038"},{url:"/fonts/caladea-v1-latin-regular.woff2",revision:"0610fc99421e0e6ca4c4c7c14ae2f43d"},{url:"/fonts/karla-v13-latin-700.eot",revision:"9c56478b8cf084983b407c96f657a2d1"},{url:"/fonts/karla-v13-latin-700.svg",revision:"a2332f3209bfe951084c20133c2e5a5d"},{url:"/fonts/karla-v13-latin-700.ttf",revision:"561294dc58294f56b0983bb5619d93da"},{url:"/fonts/karla-v13-latin-700.woff",revision:"a44f3c2705f046cd0afac4bfc2fe1eed"},{url:"/fonts/karla-v13-latin-700.woff2",revision:"57214306eebeea47c94e695a778c3f7b"},{url:"/fonts/karla-v13-latin-regular.eot",revision:"84572d5ccb359b1ae5481312e8db43ab"},{url:"/fonts/karla-v13-latin-regular.svg",revision:"581de7ca06f14fd645f791c8a283be71"},{url:"/fonts/karla-v13-latin-regular.ttf",revision:"61c4f9d8015142b069ab5e7f838594b4"},{url:"/fonts/karla-v13-latin-regular.woff",revision:"1ba33869ba2e0b430fb9cdc948dd0d54"},{url:"/fonts/karla-v13-latin-regular.woff2",revision:"c4aa4ca22f0c3e52b9e4593f63ca57a7"},{url:"/fonts/porpara-regular.ttf",revision:"49b774e03bef78a8bf8942a43fd363c9"},{url:"/icons/arrow.svg",revision:"c80dbb2c7c8ef269eb1da98c43b932f9"},{url:"/icons/back-black.svg",revision:"ab5a3aa611e11274a50ef988d38b09e5"},{url:"/icons/clock.svg",revision:"abde8858933f38dbca2c682c8499df80"},{url:"/icons/email.svg",revision:"c721c9873dcdb77b8b855ee0081a1680"},{url:"/icons/exit-white.svg",revision:"8487fd73fa6cb6e7fbf687fd9a845a50"},{url:"/icons/forwardArrow.svg",revision:"e230f09c9e2bc1337ef785dfbbcee2a2"},{url:"/icons/heart-black.svg",revision:"58370e8d994fd4619735f587e1a6e1ba"},{url:"/icons/icon-128x128.png",revision:"db7481a04968b3ad4716436d30b3e409"},{url:"/icons/icon-144x144.png",revision:"689a1926530ed6ca26db537029c1dfd7"},{url:"/icons/icon-152x152.png",revision:"98af072da209b3b297308717b06ea86c"},{url:"/icons/icon-192x192.png",revision:"7ab8c82581552c7558c5b009ffcb3cd2"},{url:"/icons/icon-384x384.png",revision:"aae6d1f552c78e06bf42aae2462a98fa"},{url:"/icons/icon-512x512.png",revision:"837ab4469c705c188326cec571ecb227"},{url:"/icons/icon-72x72.png",revision:"cb6e3fb9db7db52f0986da1ff9bab188"},{url:"/icons/icon-96x96.png",revision:"7d8348b60cdd985bd24b48cb56855c1a"},{url:"/icons/info.svg",revision:"bc2c7b826d1d41884dcd2e224d33f01b"},{url:"/icons/phone.svg",revision:"6854b54332adb9d004a730dc6fecd95e"},{url:"/icons/right-arrow.svg",revision:"1de339a612f220fda976721291a7a5ee"},{url:"/icons/url.svg",revision:"de9d153bbe15ab7d7121eef87bf12ca8"},{url:"/manifest.json",revision:"688128be216cc2f753fe641590f2fcd3"},{url:"/site.webmanifest",revision:"86e7fdaa2bd5af2341bbe0f006f05576"},{url:"/splashscreens/ipad_splash.png",revision:"e6c7e0d2a3844d883da464f677a12d6e"},{url:"/splashscreens/ipadpro1_splash.png",revision:"7e204be98022af11d093216c03bfd195"},{url:"/splashscreens/ipadpro2_splash.png",revision:"e01a2c329ec7ca63d764d53cf22f5f44"},{url:"/splashscreens/ipadpro3_splash.png",revision:"edc1e3e4677f1ce9bc224fd89979d939"},{url:"/splashscreens/iphone5_splash.png",revision:"34395737a60408ddf315ff5a98068c4e"},{url:"/splashscreens/iphone6_splash.png",revision:"aefde6cf0aa75a8ae04835884a9ea06d"},{url:"/splashscreens/iphoneplus_splash.png",revision:"12d0290aacf2a89ede283f8b0f59089e"},{url:"/splashscreens/iphonex_splash.png",revision:"f793bcec4137085c9d4ef661ef61cc91"},{url:"/splashscreens/iphonexr_splash.png",revision:"a8ec20665c9e48ffa9368faabdef6bf8"},{url:"/splashscreens/iphonexsmax_splash.png",revision:"0b914b577d2d36983fb7034dcc9b091e"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute(/^https:\/\/fonts\.(?:googleapis|gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/^https:\/\/use\.fontawesome\.com\/releases\/.*/i,new e.CacheFirst({cacheName:"font-awesome",plugins:[new e.ExpirationPlugin({maxEntries:1,maxAgeSeconds:31536e3,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.StaleWhileRevalidate({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET"),e.registerRoute(/.*/i,new e.StaleWhileRevalidate({cacheName:"others",plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400,purgeOnQuotaError:!0})]}),"GET")}));
